feat: default storage manager in main export
Added a default storage manager to main export, as well as a config() method for the StorageManager class in order to reconfigure the default storage manager without reinstantiating. This allows callers to do storage manager setup once at application start and reuse the same instances without needing to require() a separate file where a configured StorageManager has been set up. This is a similar pattern to how popular database and logging libraries export default containers for connections or loggers. Seeing as storage drivers fall under a similar class of reusability throughout an application, use of this pattern seems appropriate in this library.

Note: upgraded @types/node to version 10.17.5 to fix build issues (kept version fixed as it was previously).